Add Chatbot components: implement chat input, messages, header, and FAQ suggestions
